requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==2.2
django-haystack==2.8.1
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
numpy==1.18.4
pandas==1.0.4
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
python-slugify==4.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
